chainalysis-flask-server/requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.8.0
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
bidict==0.21.3
cbpro==1.1.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
dateparser==1.1.0
dnspython==2.1.0
eventlet==0.32.0
Flask==2.0.2
Flask-SocketIO==5.1.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
multidict==5.2.0
pymongo==3.5.1
python-binance==1.0.15
python-dateutil==2.8.2
python-engineio==4.2.1
python-socketio==5.4.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.10.23
requests==2.13.0
six==1.10.0
sortedcontainers==2.4.0
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==2.1
ujson==4.2.0
urllib3==1.26.7
websocket-client==0.40.0
websockets==9.1
Werkzeug==2.0.2
yarl==1.7.0